separate GL and GLES shader with GLES set to 1.0 for support to older versions
Also are the new coming to Kodi headers changes.
