+++ screensaver.biogenesis/addon.xml.in
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <addon
   id="screensaver.biogenesis"
-  version="2.2.0"
+  version="2.2.1"
   name="BioGenesis"
   provider-name="Team Kodi">
   <requires>@ADDON_DEPENDS@</requires>
