+++ screensaver.biogenesis/resources/shaders/GL/frag.glsl
@@ -1,0 +1,9 @@
+#version 130
+
+// Varyings
+varying vec4 v_color;
+
+void main()
+{
+  gl_FragColor = v_color;
+}
