+++ screensaver.biogenesis/resources/shaders/GL/vert.glsl
@@ -1,0 +1,14 @@
+#version 130
+
+// Attributes
+attribute vec4 a_position;
+attribute vec4 a_color;
+
+// Varyings
+varying vec4 v_color;
+
+void main()
+{
+  gl_Position = a_position;
+  v_color = a_color;
+}
