+++ screensaver.biogenesis/resources/shaders/GLES/frag.glsl
@@ -1,13 +1,11 @@
-#version 130
+#version 100
 
-#ifdef GL_ES
 precision mediump float;
-#endif
 
 // Varyings
 varying vec4 v_color;
 
-void main ()
+void main()
 {
   gl_FragColor = v_color;
 }
