+++ screensaver.biogenesis/resources/shaders/GLES/vert.glsl
@@ -1,8 +1,6 @@
-#version 130
+#version 100
 
-#ifdef GL_ES
 precision mediump float;
-#endif
 
 // Attributes
 attribute vec4 a_position;
@@ -11,7 +9,7 @@ attribute vec4 a_color;
 // Varyings
 varying vec4 v_color;
 
-void main ()
+void main()
 {
   gl_Position = a_position;
   v_color = a_color;
